Add example how to use result of ecmcConfig(OrDie) command.
The result of each ecmcConfig(OrDie) is stored in the EPICS
environment variable "ECMC_CONFIG_RETURN_VAL".

This example shows how to apply alias to records based on a
pdo value read from the EtherCAT bus (in realtime).

This example more specifically shows how to read the ID switch pdo
of an EK1101 and set an record alias based on this value, in this case
used as a crate number/identifier.

Same concept could be used for data returned from all ecmc
"ASCII" commands avaialble for ecmcConfig(OrDie).