simplify jinja2 syntax to prepare for new configuration tool
this commit aims to simplify the Jinja2 syntax used in template files to
prepare for a new configuration tool called ECB. Since ECB only supports a
subset of the Jinja2 syntax, changes have been made so, that the syntax in the
templates is simplified while still both tools can parse it.

The changes are work in progress. Open points are:

+ “epicsMotorRecord.jinja”: needs further investigation regarding homing.

+ the Python Jinja2 version seems not to parse 'axis.jinja' when axis.type=1.
  This needs more investigations.

+ all *.cmd scripts need to be adapted to switch between ECB and Jinja2.

+ More testing is needed to compare the rendered templates when both tools work
  on the changes of this commit. So far only ECB with changes of this commit is
  compared with Jinja2 and an unmodified ecmccfg repo.